Updated MoneyManager
the code should work (Unity didn't see a issue) but when i kept editing the code to add the command, there were issues so i had to compare my code with Allen's to see where the issue is, so for now I commented out Allen's code and left in my version. I also left in Allen's version so you can see side-by-side.

Co-Authored-By: NellaCu <104611548+NellaCu@users.noreply.github.com>